changed most long doubles to doubles

+++ quadrature.c
@@ -55,7 +55,7 @@ double left_sum(float a, float b, int sub_intervals) {
 
 double right_sum(float a, float b, int sub_intervals) {
     double dx = (b-a)/sub_intervals;
-    long double sum;
+    double sum;
     for(int i=0; i<sub_intervals; i++) {
         sum += f(a+(dx*(i+1)));
     }
@@ -64,7 +64,7 @@ double right_sum(float a, float b, int sub_intervals) {
 
 double midpoint_sum(float a, float b, int sub_intervals) {
     double dx = (b-a)/sub_intervals;
-    long double sum;
+    double sum;
     for(int i=0; i<sub_intervals; i++) {
         sum += f(a+(dx*i)+(dx/2));
     }
@@ -74,7 +74,7 @@ double midpoint_sum(float a, float b, int sub_intervals) {
 double trapezoidal_rule(float a, float b, int sub_intervals) {
     double dx = (b-a)/sub_intervals;
     double last_height = f(a);
-    long double sum;
+    double sum;
     for(int i=0; i<sub_intervals; i++) {
         double right_height = f(a+(dx*(i+1)));
         sum += (last_height+right_height)/2;
@@ -93,7 +93,7 @@ double simpsons_rule(float a, float b, int n) {
     return dx/3 * sum;
 }
 
-
+/*
 double simpsons_38rule(float a, float b, int sub_intervals) {
     double dx = (b-a)/sub_intervals;
     double sum;
